requirements.txt
# Kafka
kafka-python>=2.0.2
confluent-kafka==2.1.1

# Data processing
pandas==2.0.0
numpy==1.24.3

# Elasticsearch
elasticsearch>=8.0.0

# Logging
loguru>=0.6.0

# Config management
pyyaml==6.0

# For testing/development
pytest==7.3.1
faker==18.10.1

# Phase 2 Dependencies - Anomaly Detection & ML
scikit-learn==1.3.0
tensorflow==2.13.0
river==0.14.0  # For online learning and drift detection
scipy==1.11.2
ipaddress==1.0.23
matplotlib==3.7.2  # For visualization
tqdm==4.65.0  # For progress bars
joblib==1.3.2  # For model serialization

# SIEM components
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.9.0
python-dateutil>=2.8.2
requests>=2.27.1

# Phase 3 Dependencies - LangChain & Response
langchain>=0.0.267
openai>=0.28.0
faiss-cpu>=1.7.4  # For vector similarity search
paramiko>=3.3.1  # For SSH operations
tiktoken>=0.5.1  # For OpenAI token counting
tenacity>=8.2.3  # For retries in API calls